bot/requirements.txt
aiogram==2.9.2
aiohttp==3.9.3
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
certifi==2024.2.2
charset-normalizer==3.3.2
dnspython==2.6.1
frozenlist==1.4.1
idna==3.6
motor==3.4.0
multidict==6.0.5
oauthlib==3.2.2
pymongo==4.6.3
pyshorteners==1.0.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
urllib3==2.2.1
yarl==1.9.4
zoomus==1.2.1